refactor(GraphQL): General cleanup and create cron job
