+++ package.json
@@ -8,15 +8,12 @@
   "dependencies": {
     "apollo-server": "^2.21.1",
     "axios": "^0.21.1",
-    "azimuth-js": "^0.21.0",
     "graphql": "^15.5.0",
+    "graphql-request": "^3.4.0",
     "lodash.get": "^4.4.2",
-    "moment": "^2.29.1",
     "node-cron": "^3.0.0",
     "pg": "^8.5.1",
     "pg-format": "^1.0.4",
-    "ssl-root-cas": "^1.3.1",
-    "urbit-ob": "^5.0.1",
-    "web3": "^1.3.4"
+    "ssl-root-cas": "^1.3.1"
   }
 }
