+++ src/resolvers/db/populateNodeStatus.js
@@ -1,0 +1,8 @@
+const { addToDB } = require('./utils/addToDB')
+
+const populateNodeStatus = async () => {
+  await addToDB('node_status', null)
+  return true
+}
+
+module.exports = { populateNodeStatus }
