+++ src/resolvers/db/populateNodeType.js
@@ -1,0 +1,8 @@
+const { addToDB } = require('./utils/addToDB')
+
+const populateNodeType = async () => {
+  await addToDB('node_type', null)
+  return true
+}
+
+module.exports = { populateNodeType }
