+++ src/resolvers/db/utils/addToDB.js
@@ -1,19 +1,9 @@
 const { Client }    = require('pg')
 const format        = require('pg-format')
-const axios         = require('axios')
-const https         = require('https')
-const ob            = require('urbit-ob')
-const ajs           = require('azimuth-js')
-const _get           = require('lodash.get')
-const azimuth       = ajs.azimuth
-const Web3          = require('web3')
-const infura        = `https://mainnet.infura.io/v3/7014111724dc4d198c82cab378fa5453`
-
-const provider      = new Web3.providers.HttpProvider(infura)
-const web3          = new Web3(provider)
+const _get          = require('lodash.get')
 
 // Later change this to just update the DB instead of delete and replace
-const addToDB = async (tableName, columns, getDataResponse) => {
+const addToDB = async (tableName, getDataResponse) => {
   
   const client = new Client()
 
@@ -129,13 +119,6 @@ const addToDB = async (tableName, columns, getDataResponse) => {
 
     // Need to update the following code to query the raw_events table instead of a GET
     insertQuery = format(`INSERT INTO %I (%s, %s, %s, %s, %s, %s, %s) VALUES`, tableName, columnsWithoutTypes[1], columnsWithoutTypes[2], columnsWithoutTypes[3], columnsWithoutTypes[4], columnsWithoutTypes[5], columnsWithoutTypes[6], columnsWithoutTypes[7])
-    
-    let contracts 
-    try {
-      contracts = await ajs.initContractsPartial(web3, ajs.azimuth.mainnet)
-    } catch (error) {
-      console.log(`initContractsPartial error: ${error}`)
-    }
 
     console.log(`getDataResponse.length: ${getDataResponse.length}`)
 
@@ -176,8 +159,6 @@ const addToDB = async (tableName, columns, getDataResponse) => {
       const node_id = point
       
       const event_type = _get(eventTypeKey, event) || 13
-      
-      // const pointNumber = parseInt(ob.patp2dec(point))
 
       let sponsor_id = null
 
@@ -211,22 +192,6 @@ const addToDB = async (tableName, columns, getDataResponse) => {
         }
       }
 
-      // If there is no address associated with an event, it can just be null
-      // Certain types of events are associated with addresses, others are not
-      // const address = getDataResponse[i][3] || null
-      // const address = 'sample_address'
-      // let addressQuery = `select (field1, field2, field3) from (select * from raw_events order by date desc limit ${i + 1}) as nested where point = '~lompex-figrud' order by date desc limit 1;`
-      // let addressResponse
-
-      // try {
-      //   addressResponse = await client
-      //     .query(addressQuery)
-      // } catch (error) {
-      //   console.log(`createTableResponse error: ${error}`)
-      //   throw error
-      // }
-
-      // address = _get(addressResponse, 'rows[0].field1') || 1
       let address
       field1.length === 42 ? address = field1 : field2.length === 42 ? address = field2 : field3.length === 42 ? address = field3 : address = null
 
@@ -244,7 +209,6 @@ const addToDB = async (tableName, columns, getDataResponse) => {
 
       continuity_number = _get(continuityNumberResponse, 'rows[0].field1') || 1
 
-      // I think revision number is 'keys' event right?
       let revision_number
       let revisionNumberQuery = `select field1 from (select * from raw_events order by date desc limit ${i + 1}) as nested where point = '${node_id}' and event = 'keys' order by date desc limit 1;`
       let revisionNumberResponse
@@ -270,7 +234,7 @@ const addToDB = async (tableName, columns, getDataResponse) => {
       let field1
       let field2
       let field3
-      console.log("🚀 ~ file: db.js ~ line 273 ~ addToDB ~ getDataResponse[i][3]", getDataResponse[i][3])
+
       if (getDataResponse[i].length > 3 && getDataResponse[i][3] !== '') {
         field1 = getDataResponse[i][3]
       } else {
@@ -286,8 +250,6 @@ const addToDB = async (tableName, columns, getDataResponse) => {
       } else {
         field3 = null
       }
-      console.log("🚀 ~ file: db.js ~ line 192 ~ addToDB ~ getDataResponse[i]", getDataResponse[i])
-      console.log(`i: ${i}`)
       
       if (i > 0) {
         insertQuery += `,`
@@ -298,7 +260,6 @@ const addToDB = async (tableName, columns, getDataResponse) => {
     }
 
     insertQuery += ';'
-    // console.log("🚀 ~ file: db.js ~ line 299 ~ addToDB ~ insertQuery", insertQuery)
   } else if (tableName === 'radar') {
 
     insertQuery = format(`INSERT INTO %I (%s, %s, %s, %s) VALUES`, tableName, 'SHIP_NAME', 'PING', 'RESULT', 'RESPONSE')
@@ -310,10 +271,7 @@ const addToDB = async (tableName, columns, getDataResponse) => {
       }
 
       if (getDataResponse[ships[i]].length === 0) {
-        // insertQuery += format(` ('%s', '-1', '-1', '-1')`, ships[i] || null)
         insertQuery += format(` ('%s', %L, %L, %L)`, ships[i] || null, null, null, null)
-        // insertQuery += format(` ('%s')`, ships[i] || null)
-        // insertQuery += format(` ('%s', 'null', 'null', 'null')`, ships[i] || null)
       } else {
         for (let j in getDataResponse[ships[i]]) {
           if (j > 0) {
@@ -327,7 +285,6 @@ const addToDB = async (tableName, columns, getDataResponse) => {
     insertQuery += ';'
   } else if (tableName === 'ping') {
     insertQuery = format(`INSERT INTO %I (%s, %s, %s, %s) VALUES`, tableName, 'NODE_ID', 'ONLINE', 'PING_TIME', 'RESPONSE_TIME')
-    // console.log("🚀 ~ file: db.js ~ line 312 ~ addToDB ~ getDataResponse", getDataResponse)
     for (let i in getDataResponse) {
       let { ship_name, ping, response } = getDataResponse[i]
       if (i < 20) {
@@ -344,8 +301,7 @@ const addToDB = async (tableName, columns, getDataResponse) => {
       } else {
         online = true
       }
-      console.log("🚀 ~ file: db.js ~ line 330 ~ addToDB ~ online", online)
-      // ping !== '-1' ? online = true : online = false
+      
       let pingTime
       let responseTime
       if (online) {
@@ -364,14 +320,13 @@ const addToDB = async (tableName, columns, getDataResponse) => {
   insertQuery = format(`INSERT INTO %I (%s) VALUES ('%s'), ('%s'), ('%s'), ('%s'), ('%s'), ('%s'), ('%s'), ('%s'), ('%s'), ('%s'), ('%s'), ('%s');`, tableName, 'EVENT_NAME', 'change_ownership', 'change_spawn_proxy', 'change_transfer_proxy', 'change_management_proxy', 'change_voting_proxy', 'activate', 'spawn', 'escape_requested', 'escape_cancelled', 'escape_accepted', 'lost_sponsor', 'broke_continuity')
   console.log("🚀 ~ file: db.js ~ line 212 ~ addToDB ~ insertQuery", insertQuery)
 } else if (tableName === 'node_type') {
-  insertQuery = format(`CREATE TYPE %I AS ENUM ('%s' ,'%s' ,'%s', '%s', '%s');`, tableName, 'galaxy', 'star', 'planet', 'comet', 'moon')
+  insertQuery = format(`DROP TYPE node_type; CREATE TYPE %I AS ENUM ('%s' ,'%s' ,'%s', '%s', '%s');`, tableName, 'galaxy', 'star', 'planet', 'comet', 'moon')
   console.log("🚀 ~ file: db.js ~ line 212 ~ addToDB ~ insertQuery", insertQuery)
 }
 
   try {
     console.log('running addDataResponse')
     const addDataResponse = await client.query(insertQuery)
-    // console.log("🚀 ~ file: db.js ~ line 106 ~ addToDB ~ addDataResponse", addDataResponse)
   } catch (error) {
     console.log(`addDataResponse error: ${error}`)
     throw error
@@ -386,251 +341,4 @@ const addToDB = async (tableName, columns, getDataResponse) => {
   } 
 }
 
-const populateRadar = async () => {
-  console.log('running populateRadar')
-  const agent = new https.Agent({  
-    rejectUnauthorized: false
-   })
-  let events
-  try {
-    // The following can be used to test against the radar fixture in case the endpoint is down
-    // Will need to copy in the fixture, which Chris has but which is not committed due to file size
-    // events = JSON.parse(fs.readFileSync('_radar.json', 'utf8'))
-
-    events = await axios.get('http://35.247.74.19:8080/~radar.json', { httpsAgent: agent })
-    events = events.data
-  } catch (error) {
-    throw error
-  }
-  await addToDB('radar', null, events)
-  return true
-}
-
-const convertDateToISO = dateToConvert => {
-  let string = dateToConvert.slice(1)
-  string = string.split('..')
-  string[0] = string[0].replace(/\./g, '-')
-  string[1] = string[1].replace(/\./g, ':') + '.000Z'
-  string = string.join('T')
-  return string
-}
-
-// pki_event
-// ***
-// event_id (pk)
-// node_id (fk) not null
-// event_type_id (fk) not null
-// sponsor_id (fk)
-// time not null
-// address
-// continuity_number
-// revision_number
-let columns = [ 'event_id', 'node_id', 'event_type_id', 'sponsor_id', 'time', 'address', 'continuity_number', 'revision_number']
-
-const populatePKIEvents = async () => {
-  console.log('running populatePKIEvents')
-  // const agent = new https.Agent({  
-  //   rejectUnauthorized: false
-  //  })
-
-  // let events
-  // try {
-  //   console.log('in populate pki events try')
-  //   events = await axios.get('https://azimuth.network/stats/events.txt', { httpsAgent: agent })
-  // } catch (error) {
-  //   throw error
-  // }
-
-  // console.log('after populate pki events try')
-  // events = events.data
-
-  // let txtColumns = events.slice(0, events.indexOf('~')).split(',')
-  // for (let i in txtColumns) {
-  //   if (txtColumns[i].includes('\n')) {
-  //     txtColumns[i] = txtColumns[i].replace('\n', '')
-  //   }
-  //   if (txtColumns[i].includes(' ')) {
-  //     txtColumns[i] = txtColumns[i].replace(' ', '')
-  //   }
-  //   txtColumns[i] = `${txtColumns[i].toUpperCase()}`
-  // }
-
-  // events = events.slice(events.indexOf('~')).split('\n')
-
-  // for (let i in events) {
-  //   let splitString = events[i]
-  //   let splitStringArray = splitString.split(',')
-  //   if (splitStringArray[splitStringArray.length - 1] === '') {
-  //     splitStringArray.pop()
-  //   }
-  //   events[i] = splitStringArray
-    
-  //   events[i][0] = convertDateToISO(events[i][0])
-  // }
-
-  const client = new Client()
-
-  try {
-    await client.connect()
-    console.log('client connected')
-  } catch (error) {
-    console.log('client connect error')
-    throw error
-  }
-
-  let rawEventsResponse
-  try {
-    const rawEventsQueryString = format(`select * from %I;`, 'raw_events')
-    rawEventsResponse = await client.query(rawEventsQueryString)
-    console.log('client connected')
-  } catch (error) {
-    console.log('client connect error')
-    throw error
-  }
-
-  const rawEventsRows = rawEventsResponse.rows
-
-  try {
-    await client.end()
-  } catch (error) {
-    console.log(`client.end() error: ${error}`)
-    throw error
-  }
-  
-  try {
-    // await addToDB('pki_events', txtColumns, rawEventsRows)
-    await addToDB('pki_events', null, rawEventsRows)
-  } catch (error) {
-    throw error
-  }
-  
-  return true
-  
-}
-
-const populateRawEvents = async () => {
-  console.log('running populateRawEvents')
-  const agent = new https.Agent({  
-    rejectUnauthorized: false
-   })
-
-  let events
-  try {
-    console.log('in populate pki events try')
-    events = await axios.get('https://azimuth.network/stats/events.txt', { httpsAgent: agent })
-  } catch (error) {
-    throw error
-  }
-
-  console.log('after populate pki events try')
-  events = events.data
-
-  let txtColumns = events.slice(0, events.indexOf('~')).split(',')
-  for (let i in txtColumns) {
-    if (txtColumns[i].includes('\n')) {
-      txtColumns[i] = txtColumns[i].replace('\n', '')
-    }
-    if (txtColumns[i].includes(' ')) {
-      txtColumns[i] = txtColumns[i].replace(' ', '')
-    }
-    txtColumns[i] = `${txtColumns[i].toUpperCase()}`
-  }
-
-  events = events.slice(events.indexOf('~')).split('\n')
-
-  let returnArr = []
-
-  for (let i in events) {
-    let splitString = events[i]
-    let splitStringArray = splitString.split(',')
-    // if (splitStringArray[splitStringArray.length - 1] === '') {
-    //   splitStringArray.pop()
-    // }
-
-    let newArr = splitStringArray.map(x => {
-      if (x === '') {
-        return null
-      } else {
-        return x
-      }
-    })
-
-    newArr[0] = convertDateToISO(newArr[0])
-    returnArr.push(newArr)
-    // events[i] = splitStringArray
-    // events[i][0] = convertDateToISO(events[i][0])
-  }
-  
-  try {
-    // await addToDB('raw_events', txtColumns, events)
-    await addToDB('raw_events', txtColumns, returnArr)
-  } catch (error) {
-    throw error
-  }
-  
-  return true
-  
-}
-
-const populatePing = async () => {
-
-  const client = new Client()
-
-  try {
-    await client.connect()
-    console.log('client connected')
-  } catch (error) {
-    console.log('client connect error')
-    throw error
-  }
-
-  let radarResponse
-  try {
-    const radarQueryString = format(`select * from %I;`, 'radar')
-    console.log("🚀 ~ file: db.js ~ line 310 ~ populatePing ~ radarQueryString", radarQueryString)
-    radarResponse = await client.query(radarQueryString)
-    console.log('client connected')
-  } catch (error) {
-    console.log('client connect error')
-    throw error
-  }
-
-  const radarRows = radarResponse.rows
-
-  try {
-    await client.end()
-  } catch (error) {
-    console.log(`client.end() error: ${error}`)
-    throw error
-  } 
-  await addToDB('ping', null, radarRows)
-  return true
-}
-
-const populateNodeStatus = async () => {
-  await addToDB('node_status', null, null)
-  return true
-}
-
-const populateEventType = async () => {
-  await addToDB('event_type', null, null)
-  return true
-}
-
-const populateNodeType = async () => {
-  await addToDB('node_type', null, null)
-  return true
-}
-
-
-const dbResolvers = {
-  populateRadar: () => populateRadar(),
-  populatePKIEvents: () => populatePKIEvents(),
-  populateRawEvents: () => populateRawEvents(),
-  populatePing: () => populatePing(),
-  populateNodeStatus: () => populateNodeStatus(),
-  populateEventType: () => populateEventType(),
-  populateNodeType: () => populateNodeType()
-}
-
-module.exports = dbResolvers
+module.exports = { addToDB }
